Problem: allocating memory for an entire tar file
When preparing a file for sending to Docker (extension files and
Dockerifle) we are allocating memory for the entire file. It may be big,
so it'll take a lot of memory and will delay sending the file to
Docker.

Solution: tar in a blocking task and stream it to Docker