Improve registry testing docs
Document the dependencies required to run them, and instructions to
install them.

Teach the `justfile` to use the `DATABASE_URL` environment variable but
to fall back on the original value if it's not set. Also export it for
all of its targets, including the new `stop-postgres` and `test` targets.

Eliminate a warning from `registry/tests/integration_test.rs` for an
unnecessary use of `clone()`.