Refactor CLI tests and get them passing on macOS
To get the tests passing, add conditionals require the `x86_64`
architecture to test `trunk install`, and to skip them on other
architectures. This allows the individual tests to pass on macOS on
Apple Silicon.

Add `README-DEV.md` to describe the dependencies required to run the CLI
tests, as well as an instruction to limit the number of concurrent
threads so as not to crush Docker (which happened regularly with macOS
Docker's default configuration).

Refactor the tests for better file handling, including:

*   Use the tempfile::TempDir crate to create working directories
    automatically deleted when a test function finishes
*   Use Path objects instead of formatting strings for output
    directories, tarballs, and manifest files.
*   Move cloned Git repositories to the self-deleting temp directories
    rather than the repository itself.
*   Use fs::read_to_string instead of `cat` to read in files

Also: remove the `-j` option to `make` in the PostGIS test Dockerfile
(`cli/tests/test_trunk_toml_dirs/postgis/Dockerfile`) to avoid
[concurrency errors] that break the build, and remove a couple of
warnings reported by `clippy`.

  [concurrency errors]: https://stat.ethz.ch/pipermail/r-devel/2022-May/081724.html