Send Base64-encoded checksum to S3
AWS S3's [object integrity] just says it needs a checksum, but not the
format. We were using `sha256::digest()`, which returned a hex string.
Digging around in docs, finally [revealed] that it needs to be
Base64-encoded.

So replace the `sha256` crate with `sha2` from the [RustCrypto project],
which provides a binary representation of a hash digest, and encode it
as appropriate with the `base64` and `hex` crates for S3 validation and
Trunk download validation, respectively.

Also pass the digest around as a byte slice and let the uploader code
handle encoding it into Base64 for the upload request.

Resolves TEM-2861.

  [object integrity]: https://docs.aws.amazon.com/AmazonS3/latest/userguide/checking-object-integrity.html
  [revealed]: https://docs.rs/aws-sdk-s3/latest/aws_sdk_s3/operation/complete_multipart_upload/struct.CompleteMultipartUploadInput.html#method.checksum_sha256
  [RustCrypto project]: https://github.com/RustCrypto/hashes