Add plv8 3.2
As of v3.2, the plv8 project no longer includes the `coffeescript` or
`livescript` procedural languages. So add a plain old `plv8` package
to replace the `plcoffee` package, which should be kept at v3.1.x and
deprecated.

Also build on `xlarge-16x16` to escape the merciless OOM killer, which
doesn't like the super RAM-intensive plv8 build.