Update remaining core extensions to Postgres 17 (#838)
Some have new versions, either from their control files or from Postgres
itself, so adjust accordingly. Simplify the build scripting to shallowly
clone the Postgres repo at the release tag, skip configuring Postgres,
and use `make USE_PGXS=1` to install to the proper locations for the
system Postgres, rather than the `/usr/local` default from configuring
Postgres, eliminating the need for `mv`s in the install command.

The exception to the above is the procedural languages, which do not
support `USE_PGXS=1`. In those cases, configure postgres but specify
the same SHAREDIR and PKGLIBDIR as the system Postgres.

This technique does require overwriting the `Makefile` for each of the
SPI extensions, but it's quite simple, a short `printf` statement for
each.