Refactor remaining core extension build scripts (#845)
As in 8d2a1e7 (#838) make a shallow clone of the Git repo and use
`USE_PGXS=1` to build and install against the system Postgres instead of
configuring the cloned source, eliminating the need to move files in the
`install_command`.

Also, run the `install_command` with `bash` instead of `sh`, for better
shell semantics. Required for lines like this to work:

```sh
rm "$(pg_config --sharedir)"/extension/plperlu{--*,.*}
```

Make minor edits to the jsonb_plperl, plperl, and pltcl build scripts so
they fire again and properly remove unwanted files.