Refactor last bits of old-style build scripting (#882)
The exceptions are `parquet_s3_fdw` and `postgresml`, which I cannot get
to build. The updated extensions are:

*   clickhouse_fdw
*   rdkit (which never built before?)

Also simplified the use of `cmake` in these extensions:

*   mobilitydb
*   pgrouting

And upgraded these extensions to:

*   rdkit 4.6.1
*   pgrouting 3.7.1

Next, make sure that pljava properly depends on `openjdk-11-jdk`.

Upgrade the build scripting in tests:

*   cli/tests/test_builders/Dockerfile.auto_explain
*   cli/tests/test_builders/Dockerfile.http
*   cli/tests/test_builders/Dockerfile.pg_stat_statements
*   cli/tests/test_pljava
*   cli/tests/test_postgresql_unit
*   cli/tests/test_postgresql_unit
*   cli/tests/test_trunk_toml_dirs/pg_cron
*   cli/tests/test_trunk_toml_dirs/postgis

The `pljava` test revealed that the pattern adopted in  the last month
to use `make install` rather than to `mv` files broke extensions
configured with a `directory` in their control files, doubling the
directory prefix. Update `prepare_sharedir_file` to notice when files
already start with the `directory` prefix, but keep the old behavior of
stripping out `extensions` for any lingering `mv` use cases.

Finally, update the README with the cleaner output and add some notes
for running tests in Docker Desktop.