Manually specify extensions to build (#890)
Force the extensions workflow to build the extensions known to have
`directory` fields in their control files, so as to fix them. Upgrade
to `trunk-test-tembo:8a3bd64-pg15` to get the fixes, made in c3154c
(#888) and f68df13 (#891), and released in 8a3bd64 (#892).

As a result of failures revealed by fixing these issues, also teach
PL/Java where to find `libjvm.so` and the workflow to preload the `anon`
module when testing postgresql_anonymizer.