Always download DuckDB DSO (#909)
Rather than letting the Git repo build it from a submodule. Otherwise
the pg_duckdb builder installs it into the Postgres lib directory, which
isn't included in `ld`'s path, so Postgres can't find it. Update to
`trunk-test-tembo:4436f19`, which includes the supported version of
`libduckdb.so`.

Also configure it to be added to `shared_preload_libraries` and needing
to restart.

Copy the same `libduckdb.so` to build `duckdb_fdw`, too, and
explicitly set its version, since it may diverge from `duckdb_fdw`'s own
version. The build is successful, but `duckdb_fdw` does not load with
the `libduckdb.so` built by `pg_duckdb`, so skip testing it for now.
Will fix once duckdb/pg_duckdb#554 or its successor resolves thing such
that both extensions can use the same DSO.