Recognize more file extensions as Markdown (#918)
Most people use `.md`, but some use `.mmd` (for  MultiMarkdown), and
GitHub itself recognizes a number of other extensions. So look at them
all to recognize Markdown.

Also fix new linting issues and the URLs for the `colnames` and
`hostname` repositories, and match GitHub URLs that start with either
`https` or `http`.