Use official libduckdb release with duckdb_fdw (#933)
Rather than copying it from the `pg_duckdb` trunk, which is customized
and breaks `duckdb_fdw`. Which is super weird, because it builds, but
then fails to load with a missing symbol error.

So install `libduckdb.so` with its proper version and run `ldconfig`,
both in the `Dockerfile` and in the test workflow. Add it as a `lib`
dependency to the `Trunk.toml`.

While at it, use the matrix Postgres version instead of a hard-coded
`15` in the workflow.