Add pgrx 0.13.0 and refactor builder-images.yml (#942)
Refactor the `builder-images.yml` pipeline to keep local copies of each
image and to push them along with the manifest in a separate job that
runs only on `main`. This allows the `pgrx` build job to run
concurrently with the `c-builder` manifest job (on `main`) and to always
build from the just-created `c-builder` image.

Use `docker/build-push-action` Instead of using explicit `docker` calls.
This simplifies the configuration a bit and properly uses `docker
buildx`.

Remove the logins to Docker Hub, which are unused.

Reformat some of the YAML to be more compact and legible.