Replace test_pljava with test_dir_field (#954)
The PL/Java build is super slow, overkill just for testing that a
directory specification installs stuff in the right place. Replace it
with a tweaked pair tarball that doesn't have to be downloaded and
contains no dependencies. It also requires no compilation, so doesn't
have to run exclusively on x86_64, although it currently fails on an
ARM64 Mac because of some CPU mismatch between bollard and Docker.