Add support for pgrx 0.13.1 (#959)
Alas, a breaking change in cargo-edit 0.13.2 prevents pgrx 0.13.0 or
0.13.1 from building. The issue (pgcentralfoundation/pgrx#2009) is fixed
in pgcentralfoundation/pgrx@7eda7cf, so pgrx 0.13.2 should be fine. So
introduce a conditional that installs pgrx 0.13.0 and 0.13.1 with
`--locked` to prevent the installation of cargo-edit 0.13.2. It also
locks all the dependencies to the versions in the lock file, which isn't
great if there are security fixes, but at least it builds, and hopefully
most people will upgrade to 0.13.2 soon after it comes out.