Release CLI on tag rather than push to main (#974)
The `cli.yml` workflow publishes to crates.io on eery every push to
main. But that doesn't work for releasing on GitHub, because GitHub
releases require a tag, and as the comment in `cli.yml` reports, pushing
a tag from a workflow does not trigger other workflows.

So change the CLI release process to require a tag. Validate that the
tag matches the version in `cli/Cargo.toml`. This means we'll need to be
diligent about tagging CLI releases, as they can't be automated in the
same way as `cargo publish`.