Remove arch from Rust cache-key (#974)
The Rust setup action already includes the OS and architecture in the
key, so no need to add it.

Increment to v0.16.8, as the v0.16.7 release included an arm64 binary in
the amd64 artifact, likely due to a cache hit for a previous incorrect
build.

Also trigger builds in pull requests targeting the main branch.